requirements.txt
asgiref==3.5.0
backports.entry-points-selectable==1.1.1
certifi==2021.10.8
colorama==0.4.4
distlib==0.3.4
Django==4.0.1
django-jazzmin==2.4.9
djangorestframework==3.13.1
filelock==3.4.0
Pillow==9.0.0
pipenv==2022.1.8
platformdirs==2.4.0
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
virtualenv==20.10.0
virtualenv-clone==0.5.7
